Feature/add response page

+++ src/aims_ui/page_multiple_address.py
@@ -1,3 +1,7 @@
+import json
+import logging
+import csv
+from time import sleep
 import os
 from flask import render_template, request, session, send_file
 from requests.exceptions import ConnectionError
@@ -13,10 +17,8 @@ from .models.get_addresses import get_addresses
 from .upload_utils import check_valid_upload
 from .page_error import page_error
 from .upload_utils import FileUploadException
-import json
-import logging
-import csv
-from time import sleep
+from .security_utils import check_user_has_access_to_page
+from .google_utils import get_username, get_current_group
 
 page_name = 'multiple_address'
 
@@ -40,6 +42,19 @@ def request_entity_too_large(error):
 @login_required
 @app.route(f'/{page_name}', methods=['GET', 'POST'])
 def multiple_address():
+  endpoints = get_endpoints(called_from=page_name)
+  access = check_user_has_access_to_page(page_name, endpoints)
+  if access != True:
+    return access
+
+  username = get_username()
+  current_group = get_current_group()
+  if username in current_group.get('usernames'):
+    reduced = True
+    limit = current_group.get('limit_mini_bulk')
+  else:
+    reduced = False
+    limit = 5000
 
   if request.method == 'GET':
     delete_input(session)
@@ -52,11 +67,11 @@ def multiple_address():
     return render_template(
         f'{page_name}.html',
         searchable_fields=searchable_fields,
-        endpoints=get_endpoints(called_from=page_name),
+        endpoints=endpoints,
+        limit=limit,
     )
 
   if request.method == 'POST':
-
     searchable_fields = get_fields(page_name)
     all_user_input = load_save_store_inputs(
         searchable_fields,
@@ -69,31 +84,35 @@ def multiple_address():
     try:
       # Validate file
       file_valid, error_description, error_title = check_valid_upload(
-          file, limit=1000000)
+          file, limit=limit)
     except FileUploadException as e:
       return final(searchable_fields,
+                   limit,
                    error_description=e.error_description,
                    error_title=e.error_title)
 
     if file_valid:
       multiple_address_match(file, all_user_input, download=True)
-      return final(all_user_input)
+      return final(all_user_input, limit)
 
 
 def final(all_user_input,
+          limit,
           error_description='',
           error_title='',
           results_summary_table='',
           table_results=''):
+  endpoints = get_endpoints(called_from=page_name)
   searchable_fields = get_fields(page_name)
 
   return render_template(
       f'{page_name}.html',
       error_description=error_description,
       error_title=error_title,
-      endpoints=get_endpoints(called_from=page_name),
+      endpoints=endpoints,
       searchable_fields=searchable_fields,
       table_results=table_results,
       results_summary_table=results_summary_table,
       results_page=True,
+      limit=limit,
   )
