+++ src/aims_ui/page_singlesearch.py
@@ -1,16 +1,16 @@
 import os
+import json
 from flask import render_template, request, session
 from flask_login import login_required
-from . import app
 from requests.exceptions import ConnectionError
+from . import app
 from .cookie_utils import save_input, load_input, get_all_inputs, delete_input, load_save_store_inputs, save_epoch_number
 from .api_interaction import api, get_response_attributes
-from .security_utils import detect_xml_injection
+from .security_utils import detect_xml_injection, check_user_has_access_to_page
 from .models.get_endpoints import get_endpoints
 from .models.get_fields import get_fields
 from .models.get_addresses import get_addresses
 from .page_error import page_error
-import json
 
 page_name = 'singlesearch'
 
@@ -18,13 +18,17 @@ page_name = 'singlesearch'
 @login_required
 @app.route(f'/', methods=['GET', 'POST'])
 def singlesearch():
+  endpoints = get_endpoints(called_from=page_name)
+  access = check_user_has_access_to_page(page_name, endpoints)
+  if access != True:
+    return access
 
   if request.method == 'GET':
     delete_input(session)
     return render_template(
         f'{page_name}.html',
         searchable_fields=get_fields(page_name),
-        endpoints=get_endpoints(called_from=page_name),
+        endpoints=endpoints,
     )
 
   searchable_fields = get_fields(page_name)
@@ -75,11 +79,11 @@ def singlesearch():
     )
 
   # Get the attributes of the Response a user might want
-  responseAttributes = get_response_attributes(result.json());
+  responseAttributes = get_response_attributes(result.json())
 
   return render_template(
       f'{page_name}.html',
-      endpoints=get_endpoints(called_from=page_name),
+      endpoints=endpoints,
       searchable_fields=searchable_fields,
       results_page=True,
       matched_addresses=matched_addresses,
