+++ src/aims_ui/page_typeahead.py
@@ -1,14 +1,15 @@
 import os
+import json
 from flask import render_template, request, session
 from flask_login import login_required
-from . import app
 from requests.exceptions import ConnectionError
+from . import app
 from .cookie_utils import save_input, load_input, get_all_inputs, delete_input, load_save_store_inputs, save_epoch_number
 from .api_interaction import api, get_api_auth
+from .security_utils import check_user_has_access_to_page
 from .models.get_endpoints import get_endpoints
 from .models.get_fields import get_fields
 from .models.get_addresses import get_addresses
-import json
 
 page_name = 'typeahead'
 
@@ -16,6 +17,10 @@ page_name = 'typeahead'
 @login_required
 @app.route(f'/{page_name}', methods=['GET', 'POST'])
 def typeahead():
+  endpoints = get_endpoints(called_from=page_name)
+  access = check_user_has_access_to_page(page_name, endpoints)
+  if access != True:
+    return access
 
   if request.method == 'GET':
     delete_input(session)
